backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
django-filter>=23.5
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3

# File handling
Pillow>=10.1
PyPDF2>=3.0
python-docx>=1.1
python-magic-bin>=0.4.14  # For Windows

# AI & ML
google-generativeai>=0.3

# Background jobs
celery>=5.3
redis>=5.0
django-celery-beat>=2.5  # Periodic tasks

# Email
django-anymail[sendgrid]>=10.2  # Optional: better email

# API documentation
drf-spectacular>=0.27  # OpenAPI/Swagger

# Security
django-environ>=0.11

# Utilities
python-dateutil>=2.8
pytz>=2023.3
